Fix PostgreSQL type casting issues in Week 9 solutions (#32)
* fix(data-track/week-9): clarify validation expectations, deterministic sample, assignment cast note

Polish from a Week 9 exercise review (all solutions live-verified against the
shared Azure Postgres: Manhattan top, 0 NULL pickups, 189 duplicate groups,
0 orphans, fact view 56,369 = 56,551 - 182 negatives).

- Ex1: add ORDER BY pickup_datetime before LIMIT 5 so the sample is stable
  between runs (a bare LIMIT returns an arbitrary 5 rows).
- Ex5: tell students 5a returns 0 and 5c returns empty BY DESIGN (pickup IDs
  are complete; orphans do not exist on this load). A 0 is the check passing,
  not a mistake. Prevents the "did I break it?" confusion.
- Ex6: note the practice view is scaled down; the assignment's vw_fact_trips
  also casts pickup_datetime::TIMESTAMP. Flag it so students do not under-build
  the assignment deliverable.


* Fix double precision to numeric type casting issue for ROUND() in week 9 exercise solutions

---------

+++ data-track/week-9/exercise_2/solutions/exercise.sql
@@ -3,7 +3,7 @@
 SELECT
     z.borough,
     COUNT(*) AS total_trips,            -- WHY COUNT(*): counts every row in the group, which is one row per trip, so this is the trip count per borough
-    ROUND(AVG(t.fare_amount), 2) AS avg_fare  -- WHY ROUND: AVG returns many decimal places; rounding to 2 keeps the fare readable as currency
+    ROUND(AVG(t.fare_amount)::numeric, 2) AS avg_fare  -- WHY ROUND: AVG returns many decimal places; rounding to 2 keeps the fare readable as currency
 FROM nyc_taxi.raw_trips t
 INNER JOIN nyc_taxi.raw_zones z
     ON t.pickup_location_id = z.location_id
